engines/physics_engines/pinocchio/python/dtack/backends/pink_backend.py
"""PINK backend wrapper for inverse kinematics."""

from __future__ import annotations

import logging
import typing
from pathlib import Path

if typing.TYPE_CHECKING:
    import numpy as np
    import numpy.typing as npt

logger = logging.getLogger(__name__)

# PINK availability check
try:
    import pink  # noqa: F401

    PINK_AVAILABLE = True
except ImportError:
    PINK_AVAILABLE = False
    logger.warning("PINK not available. Install with: pip install pink")


class PINKBackend:
    """PINK backend for inverse kinematics.

    This backend provides:
    - IK task definition
    - Closed-loop IK solving
    - Task-space control
    """

    def __init__(self, model_path: Path | str) -> None:
        """Initialize PINK backend.

        Args:
            model_path: Path to model file or canonical YAML specification

        Raises:
            ImportError: If PINK is not installed
        """
        if not PINK_AVAILABLE:
            msg = "PINK is required but not installed. Install with: pip install pink"
            raise ImportError(msg)

        self.model_path = Path(model_path)
        logger.info("PINK backend initialized (stub implementation)")
        # NOTE: Implement PINK model loading

    def solve_ik(
        self,
        _tasks: dict[str, npt.NDArray[np.float64]],
        _q_init: npt.NDArray[np.float64],
    ) -> npt.NDArray[np.float64]:
        """Solve inverse kinematics for given tasks.

        Args:
            tasks: Dictionary of task names to target poses/positions
            q_init: Initial joint configuration

        Returns:
            Joint configuration satisfying tasks
        """
        # NOTE: Implement PINK IK solver
        msg = "PINK IK solver not yet implemented"
        raise NotImplementedError(msg)
