engines/physics_engines/pinocchio/python/requirements.txt
# Core scientific computing libraries
numpy==2.2.6
pandas==2.3.3
matplotlib==3.9.1
scipy==1.15.3

# Testing and development
pytest==8.2.1
pytest-cov==7.0.0
pytest-mock==3.15.1

# Code quality and formatting
ruff==0.5.0
black==25.12.0
mypy==1.19.0
pre-commit==4.5.0

# Configuration and utilities
pyyaml==6.0.1
pathlib2==2.3.7.post1

# Type stubs for better type checking
types-PyYAML==6.0.12.20250915

# ============================================================================
# Modeling Stack (Pinocchio, Pink, Drake, QP solvers)
# ============================================================================
# Pinocchio - Analytical dynamics library
# Use specific version to avoid dependency resolution depth issues
# Note: 'pin' package on PyPI is unrelated.
# For conda, use 'pinocchio'. For pip, use 'libpinocchio'.
# This requirements file lists 'pinocchio' for conda environments only.
pinocchio>=2.6.17

# Pink (pin-pink) - Stacked on top of Pinocchio for advanced kinematics
# Requires pin, so install after pin
pin-pink>=0.1.0,<4.0.0

# Drake - Robot manipulation and control
# Large dependency tree - constrain to stable version
drake>=1.0.0,<2.0.0

# QP Solvers - Quadratic programming solvers
# Add version constraints to help pip resolve dependencies
# Note: qpsolvers 1.x maxes out at 1.10.0 (no 1.12.0 exists)
qpsolvers>=1.10.0,<5.0.0
osqp>=0.6.0,<2.0.0
scs>=3.2.0,<4.0.0
quadprog>=0.1.0,<1.0.0

# ============================================================================
# Double Pendulum Model Dependencies
# ============================================================================
sympy>=1.12
PyQt6>=6.7
meshcat>=0.0.18
mujoco>=3.0.0